requirements.txt
Pillow
tkinterdnd2-universal
ttkthemes
